requirements.txt
asgiref==3.8.1
Django==4.2.13
django-environ==0.11.2
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
djangorestframework
psycopg2-binary
pillow
drf-yasg
pyyaml 
uritemplate 
inflection
django-filter
djangorestframework-simplejwt